refactor(be): 컨트롤러 분리 및 SSE 사설정 삭제
